Move editAdvance to editQuestion
