--- web/portal/src/components/EditPanel/EditAdvance/style.css
@@ -1,17 +1,0 @@
-/*
- * Style for components/EditPanel/EditAdvance/index.js
- */
-
-.item {
-    label {
-        font-weight: normal;
-    }
-
-    input[type="checkbox"] {
-        margin-right: 10px;
-    }
-
-    input[type="text"] {
-        width: 140px;
-    }
-}
