+++ web/portal/src/components/EditPanel/EditQuestion/style.css
@@ -24,4 +24,18 @@
         width: 360px;
         resize: none;
     }
+
+    .item {
+        label {
+            font-weight: normal;
+        }
+
+        input[type="checkbox"] {
+            margin-right: 10px;
+        }
+
+        input[type="text"] {
+            width: 140px;
+        }
+    }
 }
